now plays video background if found, not including it cuz its goddamn 500mb. bullet option also expanded

+++ core/src/com/god/fractal/Bullet.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.physics.box2d.Shape;
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
 
 public abstract class Bullet extends Entity{
-    public int dmg; //damage of the bullet
-    public double lifeTime;
+    public float dmg; //damage of the bullet
+    public float lifeTime;
 
 }
